# User/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Este API REST se encarga de gestionar la información y operaciones relacionadas con los usuarios dentro del microservicio de Usuarios del sistema de Netflix. Proporciona diversos endpoints para manejar el ciclo de vida completo de los usuarios, incluyendo la creación de cuentas, autenticación, actualización de perfiles, y la gestión de preferencias y configuraciones de cada usuario. Permite realizar operaciones CRUD (Crear, Leer, Actualizar, Eliminar) sobre los datos de usuario, asegurando que la información esté siempre segura y actualizada. Esta API facilita la interacción entre el frontend y el backend del sistema, garantizando que los usuarios puedan gestionar fácilmente sus cuentas y preferencias, mejorando la experiencia personalizada y segura dentro de la plataforma.
  title: API de Usuario
  version: "1.0"
servers:
- url: https://api.server/v1/user/
...
paths:
  /users:
    get:
      description: Obtiene todos los usuarios registrados en la plataforma.
      operationId: getUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: Operación exitosa
        "400":
          description: Solicitud inválida
        "404":
          description: Usuarios no encontrados
      summary: Obtiene todos los usuarios
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
...
components:
  schemas:
    User:
      example:
        createdAt: 2023-10-09T10:20:30Z
        password: 1e34r
        preferences:
        - 1
        - 3
        - 4
        name: Juan Pérez
        userID: 1
        email: juan.perez@example.com
        username: juanperez
      properties:
        userID:
          description: ID único del usuario
          example: 1
          type: integer
        username:
          description: Nombre de usuario
          example: juanperez
          type: string
        password:
          description: Contraseña
          example: 1e34r
          type: string
        email:
          description: Dirección de correo electrónico del usuario
          example: juan.perez@example.com
          format: email
          type: string
        name:
          description: Nombre completo del usuario
          example: Juan Pérez
          type: string
        createdAt:
          description: Fecha y hora en que se creó el usuario
          example: 2023-10-09T10:20:30Z
          format: date-time
          type: string
        preferences:
          description: Array con los id's de géneros de contenido que le gustan al
            usuario
          example:
          - 1
          - 3
          - 4
          items:
            type: integer
          type: array
      required:
      - createdAt
      - email
      - name
      - preferences
      - userID
      - username
      type: object
...